+++ packages/framework/src/Concerns/HydePage.php
@@ -31,6 +31,7 @@ abstract class HydePage implements CompilableContract, PageSchema
     use ConstructsPageSchemas;
     use Internal\HandlesPageFilesystem;
     use Internal\HandlesPageRouting;
+    use Internal\HandlesPageMatter;
 
     public static string $sourceDirectory;
     public static string $outputDirectory;
@@ -88,38 +89,6 @@ abstract class HydePage implements CompilableContract, PageSchema
     public static function all(): PageCollection
     {
         return Hyde::pages()->getPages(static::class);
-    }
-
-    /**
-     * Get a value from the computed page data, or fallback to the page's front matter, then to the default value.
-     *
-     * @return \Hyde\Framework\Models\FrontMatter|mixed
-     */
-    public function get(string $key = null, mixed $default = null): mixed
-    {
-        if ($key !== null && property_exists($this, $key) && isset($this->$key)) {
-            return $this->$key;
-        }
-
-        return $this->matter($key, $default);
-    }
-
-    /**
-     * Get the front matter object, or a value from within.
-     *
-     * @return \Hyde\Framework\Models\FrontMatter|mixed
-     */
-    public function matter(string $key = null, mixed $default = null): mixed
-    {
-        return $this->matter->get($key, $default);
-    }
-
-    /**
-     * See if a value exists in the computed page data or the front matter.
-     */
-    public function has(string $key): bool
-    {
-        return ! blank($this->get($key));
     }
 
     /**
